odyssey-planex(backend-seam): EXECUTE — T3 拆出 Pi subprocess attempt
execution.ts 2798 → 966 行(纯编排)，attempt 实现 1949 行独立成
runs/pi-subprocess-attempt.ts。同目录搬移，46 个相对导入路径零改动。

分层修正：原计划把 execution-infra.ts(2421 行)迁入 pi-maestro-backends，
但它 import agents/routing/session-handoff/model-routing/child-extensions
五个 teammate 内部值模块——迁过去只是把循环依赖换个方向。这些是所有后端
共用的编排概念，本就该留在 teammate。改为在包内拆分：

  backend-core(契约) ← backends(registry/config/裁决，不认识 Pi) ← teammate(编排 + Pi 实现)
  第三方 backend(dsh) 只依赖 backend-core

execution.ts 保留候选扫描、熔断、重放围栏、completion 发布——决定哪个模型跑；
attempt 模块拥有某个具体运行时怎么跑它。恢复 WeakMap 降为模块私有：写者与
读者现在同处一文件。

先前公开的 5 个值 + 1 个类型原样 re-export，消费方 import 路径一处未动。

两个源码正则守护测试改为拼接读取两个模块——它们守护的是派发路径的不变量
(生命周期归属、参数脱敏)，不是文件名。行为未变。

测试 1306/1304/2，与上游干净基线 9ee5edbe 完全一致。
